requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.4.0
httpx==0.27.2
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.7
python-dotenv==1.0.1
orjson==3.10.7
PyJWT==2.9.0 
google-auth==2.32.0
requests==2.32.3